[Refactor][6/N] Deepseek and Qwen merged base
