--- vllm_ascend/spec_decode/_dspark_proposer.py
@@ -1,388 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-
-from typing import Any
-
-import torch
-from vllm.config import CUDAGraphMode, VllmConfig, get_layers_from_vllm_config
-from vllm.forward_context import get_forward_context
-from vllm.model_executor.layers.attention_layer_base import AttentionLayerBase
-from vllm.v1.attention.backends.utils import CommonAttentionMetadata
-from vllm.v1.kv_cache_interface import UniformTypeKVCacheSpecs
-from vllm.v1.worker.utils import AttentionGroup
-
-from vllm_ascend.ascend_forward_context import _EXTRA_CTX, set_ascend_forward_context
-from vllm_ascend.attention.attention_v1 import AscendAttentionState
-from vllm_ascend.attention.utils import AscendCommonAttentionMetadata
-from vllm_ascend.spec_decode.dflash_proposer import AscendDflashProposer
-from vllm_ascend.ops.triton.spec_decode.utils import copy_and_expand_dflash_and_dspark_inputs_kernel_single_grid
-
-
-class AscendDSparkProposer(AscendDflashProposer):
-    """DSpark block proposer.
-
-    DSpark uses vLLM's ``mtp`` method in user config, but its execution shape is
-    closer to DFlash: target hidden states prepopulate draft K/V, then one
-    anchor-first query block emits all speculative tokens.
-    """
-
-    def __init__(
-        self,
-        vllm_config: VllmConfig,
-        device: torch.device,
-        runner=None,
-    ):
-        super().__init__(vllm_config, device, runner=runner)
-        assert vllm_config.speculative_config is not None
-        if vllm_config.speculative_config.draft_sample_method == "probabilistic":
-            raise ValueError(
-                "DSpark probabilistic draft sampling is not supported on the v1 "
-                "model runner; use greedy (the default) instead."
-            )
-        self._dspark_seed_buffer = torch.zeros(self.max_batch_size, dtype=torch.int64, device=device)
-        # DSpark is not supported in vllm v1, so related property needs to be reset here.
-        self.hidden_size = vllm_config.speculative_config.draft_model_config.get_hidden_size()
-        self.hidden_states = torch.zeros(
-            (self.max_num_tokens, self.hidden_size),
-            dtype=self.dtype,
-            device=self.device,
-        )
-        self._dflash_hidden_states = torch.zeros(
-            (self.max_num_tokens, self.hidden_size),
-            dtype=self.dtype,
-            device=self.device,
-        )
-        # DSpark runs eager only (Ascend cudagraph unsupported on this path).
-        self.use_cuda_graph = False
-        # Max query tokens = max_batch_size * num_speculative_tokens
-        # (anchor-first: N query tokens per request, no bonus token, unlike
-        # DFlash's 1+N). Overrides dflash:28; v2 derives via num_query_per_req.
-        self.max_query_tokens = self.max_batch_size * self.num_speculative_tokens
-        # Position ids for the draft query block [max_query_tokens].
-        # Overrides dflash:49; v2 uses input_buffers.positions.
-        self.positions = torch.zeros(
-            self.max_query_tokens,
-            dtype=torch.int32,
-            device=device,
-        )
-        # Primary-group query slot mapping buffer [max_query_tokens].
-        # Overrides dflash:37; v2 uses BlockTables.slot_mappings. Per-non-
-        # primary-gid buffers live in _per_group_query_slot_mapping_buffers.
-        self._slot_mapping_buffer = torch.zeros(
-            self.max_query_tokens,
-            dtype=torch.int32,
-            device=device,
-        )
-        # Cached slices of the runner's common_attn_metadata (cad): in
-        # set_inputs_first_pass they alias cad.query_start_loc_cpu / cad.seq_lens;
-        # in dummy_run/profile_run (no cad available) they are synthesized locally.
-        # Kept under the _dummy_ prefix because llm_base already defines
-        # self.query_start_loc for Eagle.
-        self._dummy_query_start_loc: torch.Tensor | None = None
-        self._dummy_seq_lens: torch.Tensor | None = None
-
-        # TODO simplify these comments
-        # block_table / slot_mapping bookkeeping (10 dicts below). v1 self-
-        # manages per kv_cache_group_id / per layer because it lacks v2's
-        # BlockTables scaffold; v2 injects a single self.block_tables
-        # (BlockTables, with .slot_mappings) + build_slot_mappings_by_layer,
-        # so the speculator holds none of these. P2 refactor target (move to
-        # runner).
-
-        # per-gid block_table from runner (just read)
-        self._per_group_block_tables: dict[int, torch.Tensor] = {}
-        # per-gid slot_mapping from runner (just read)
-        self._per_group_slot_mappings: dict[int, torch.Tensor] = {}
-
-        # per-gid block_table (use in proposer)
-        self._per_group_block_table_buffers: dict[int, torch.Tensor] = {}
-        # per-gid query slot_mapping buffer
-        self._per_group_query_slot_mapping_buffers: dict[int, torch.Tensor] = {}
-        # per-gid context slot_mapping buffer
-        self._per_group_context_slot_mapping_buffers: dict[int, torch.Tensor] = {}
-
-        # per-layer context slot mappings as a flat list
-        self._context_slots: list[torch.Tensor | None] = None
-
-
-    def initialize_attn_backend(self, kv_cache_config, kernel_block_sizes=None) -> None:
-        # Find draft layers (attention layers added by draft model)
-        all_attn_layers = get_layers_from_vllm_config(
-            self.vllm_config,
-            AttentionLayerBase,  # type: ignore[type-abstract]
-        )
-
-        attention_groups_list: list[dict[tuple[str, str], AttentionGroup]] = []
-        # the draft layers have multiple kv_cahce_groups
-        if not hasattr(self.model, "get_draft_kv_cache_layer_names"):
-            raise RuntimeError(
-                "DSpark standard-cache path requires the draft model to expose "
-                "get_draft_kv_cache_layer_names"
-            )
-
-        self._draft_attn_layer_names = sorted(self.model.get_draft_kv_cache_layer_names())
-
-        # there are many kv groups other than one
-        for kv_cache_gid, kv_cache_group_spec in enumerate(kv_cache_config.kv_cache_groups):
-            draft_layer_names_in_group = set(kv_cache_group_spec.layer_names) & set(self._draft_attn_layer_names)
-            if not draft_layer_names_in_group:
-                continue
-
-            attention_groups: dict[tuple[str, Any], AttentionGroup] = {}
-            # iterate in a way like vllm's llm_base_proposer
-            for layer_name in draft_layer_names_in_group:
-                attn_backend = all_attn_layers[layer_name].get_attn_backend()
-                layer_kv_cache_spec = kv_cache_group_spec.kv_cache_spec
-                if isinstance(layer_kv_cache_spec, UniformTypeKVCacheSpecs):
-                    layer_kv_cache_spec = layer_kv_cache_spec.kv_cache_specs[layer_name]
-                key = (attn_backend.full_cls_name(), layer_kv_cache_spec)
-                
-                if key not in attention_groups:
-                    attn_group = AttentionGroup(
-                        attn_backend,
-                        [layer_name],
-                        layer_kv_cache_spec,
-                        kv_cache_gid,
-                    )
-                    attn_group.create_metadata_builders(self.vllm_config, self.device)
-                    attention_groups[key] = attn_group
-                else:
-                    attention_groups[key].layer_names.append(layer_name)
-
-            attention_groups_list.append(attention_groups)
-
-        self.draft_attn_groups = [attention_group for attention_groups in attention_groups_list for attention_group in attention_groups.values()]
-        self.kv_cache_gid = 0
-        if not self.draft_attn_groups:
-            raise RuntimeError(
-                "DSpark standard-cache path requires registered draft attention "
-                f"groups. Missing layers: {self._draft_attn_layer_names}"
-            )
-
-        self.kv_cache_gid = self.draft_attn_groups[0].kv_cache_group_id
-        self.kernel_block_size = int(self.draft_attn_groups[0].kv_cache_spec.block_size)
-
-        name_to_gid = {
-            ln: gid
-            for gid, group in enumerate(kv_cache_config.kv_cache_groups)
-            for ln in group.layer_names if ln in self._draft_attn_layer_names
-        }
-        self._layer_group_idx = [name_to_gid[name] for name in self._draft_attn_layer_names]
-
-        # some buffers need information of groups
-        self._per_group_query_slot_mapping_buffers = {attn_group.kv_cache_group_id: torch.zeros(self.max_query_tokens, dtype=torch.int32, device=self.device) for attn_group in self.draft_attn_groups}
-        self._per_group_context_slot_mapping_buffers = {attn_group.kv_cache_group_id: torch.zeros(self.max_num_tokens, dtype=torch.int32, device=self.device) for attn_group in self.draft_attn_groups}
-
-
-    def set_per_group_attn_metadata(
-        self,
-        gid: int,
-        block_table: torch.Tensor,
-        slot_mapping: torch.Tensor,
-    ) -> None:
-        self._per_group_block_tables[gid] = block_table
-        self._per_group_slot_mappings[gid] = slot_mapping
-
-
-    def set_inputs_first_pass(
-        self,
-        target_token_ids: torch.Tensor,
-        next_token_ids: torch.Tensor,
-        target_positions: torch.Tensor,
-        target_hidden_states: torch.Tensor,
-        token_indices_to_sample: torch.Tensor | None,
-        cad: CommonAttentionMetadata,
-        num_rejected_tokens_gpu: torch.Tensor | None,
-        req_scheduled_tokens=None,
-        long_seq_metadata=None,
-        num_prefill_reqs=0,
-        num_decode_reqs=0,
-    ) -> tuple[int, torch.Tensor, CommonAttentionMetadata, tuple[Any, Any] | None]:
-        # The initial input token of markovHead is the next token
-        n = next_token_ids.shape[0]
-        self._dspark_seed_buffer[:n].copy_(next_token_ids)
-        self._dspark_seed_buffer[n:].fill_(0)
-        batch_size = cad.num_reqs
-        block_size = self.num_speculative_tokens
-        num_query_total = batch_size * block_size
-        has_num_rejected = num_rejected_tokens_gpu is not None
-        primary_gid = getattr(self, "kv_cache_gid", 0)
-        per_group_draft_block_tables = {attn_group.kv_cache_group_id: self._per_group_block_tables[attn_group.kv_cache_group_id] for attn_group in self.draft_attn_groups}
-        self._per_group_block_table_buffers = per_group_draft_block_tables
-        self._context_slots = None
-        self._dflash_num_context = int(cad.query_start_loc_cpu[batch_size])
-        self._dflash_hidden_states[:self._dflash_num_context] = target_hidden_states[:self._dflash_num_context]
-
-        # below (SAMPLE_FROM_ANCHOR=True, anchor included) -- not arange here.
-        token_indices_to_sample = torch.empty(
-            num_query_total,
-            dtype=torch.int32,
-            device=self.device,
-        )
-
-        # Query block: reuse the DFlash inputs kernel logic (host-side ref)
-        # per kv-cache-group to fill positions / input_ids / query slot_mapping
-        # / token_indices (SAMPLE_FROM_ANCHOR: anchor at q_idx=0 is sampled too).
-        draft_attn_groups = getattr(self, "draft_attn_groups", [])
-        if per_group_draft_block_tables and draft_attn_groups:
-            for attn_group in draft_attn_groups:
-                gid = attn_group.kv_cache_group_id
-                gid_block_table = per_group_draft_block_tables.get(gid)
-                if gid_block_table is None:
-                    continue
-                kv_block_size = int(attn_group.kv_cache_spec.block_size)
-                copy_and_expand_dflash_and_dspark_inputs_kernel_single_grid[1,](
-                    # Inputs
-                    next_token_ids_ptr=next_token_ids,
-                    target_positions_ptr=target_positions,
-                    context_slot_mapping_ptr=self._per_group_slot_mappings[gid],
-                    # Outputs
-                    out_input_ids_ptr=self.input_ids,
-                    out_context_positions_ptr=self._context_positions_buffer,
-                    out_query_positions_ptr=self.positions,
-                    out_context_slot_mapping_ptr=self._per_group_context_slot_mapping_buffers[gid],
-                    out_query_slot_mapping_ptr=self._per_group_query_slot_mapping_buffers[gid],
-                    out_token_indices_ptr=token_indices_to_sample,
-                    # Block table
-                    block_table_ptr=gid_block_table,
-                    block_table_stride=gid_block_table.stride(0),
-                    # Metadata
-                    query_start_loc_ptr=cad.query_start_loc,
-                    seq_lens_ptr=cad.seq_lens,
-                    num_rejected_tokens_ptr=num_rejected_tokens_gpu,
-                    # Scalars
-                    parallel_drafting_token_id=self.parallel_drafting_token_id,
-                    block_size=kv_block_size,
-                    num_query_per_req=block_size,
-                    num_speculative_tokens=block_size,
-                    total_input_tokens=self._dflash_num_context,
-                    batch_size=batch_size,
-                    HAS_NUM_REJECTED=has_num_rejected,
-                    SAMPLE_FROM_ANCHOR=True,
-                )
-        else:
-            # No draft attn groups (profile/dummy without standard path):
-            # fill positions/input_ids only.
-            for req_idx in range(batch_size):
-                ctx_end = int(cad.query_start_loc[req_idx + 1].item())
-                valid_ctx_end = ctx_end
-                if has_num_rejected:
-                    assert num_rejected_tokens_gpu is not None
-                    valid_ctx_end -= int(num_rejected_tokens_gpu[req_idx].item())
-                last_pos = target_positions[valid_ctx_end - 1]
-                out_start = req_idx * block_size
-                out_end = out_start + block_size
-                self.positions[out_start:out_end] = last_pos + 1 + self.arange_dflash[:block_size]
-                self.input_ids[out_start] = next_token_ids[req_idx]
-                if block_size > 1:
-                    self.input_ids[out_start + 1 : out_end] = self.parallel_drafting_token_id
-
-        # to compute self._context_slots
-        if per_group_draft_block_tables:
-            per_group_draft_context_slot_mappings = {
-                gid: self._per_group_context_slot_mapping_buffers[gid][:self._dflash_num_context]
-                for gid in per_group_draft_block_tables
-            }
-            self._context_slots = [per_group_draft_context_slot_mappings[gidx] for gidx in self._layer_group_idx]
-
-        effective_seq_lens = cad.seq_lens
-        if has_num_rejected:
-            effective_seq_lens = effective_seq_lens - num_rejected_tokens_gpu
-
-        cad.query_start_loc = self.arange_dflash[: batch_size + 1] * block_size
-        cad.seq_lens = effective_seq_lens + block_size
-        cad.query_start_loc_cpu = (torch.from_numpy(self.token_arange_np[: batch_size + 1]).clone() * block_size).to(
-            torch.int32
-        )
-
-        if hasattr(cad, "actual_seq_lengths_q"):
-            cad.actual_seq_lengths_q = [block_size] * batch_size
-        if hasattr(cad, "decode_token_per_req"):
-            cad.decode_token_per_req = block_size
-
-        cad.num_actual_tokens = num_query_total
-        cad.num_input_tokens = num_query_total
-        cad.max_query_len = block_size
-        cad.max_seq_len = cad.max_seq_len + block_size
-        cad.slot_mapping = self._slot_mapping_buffer[:num_query_total]
-        if per_group_draft_block_tables:
-            self._per_group_query_slot_mapping_buffers = {
-                gid: self._per_group_query_slot_mapping_buffers[gid] for gid in per_group_draft_block_tables
-            }
-            if primary_gid in self._per_group_query_slot_mapping_buffers:
-                cad.slot_mapping = self._per_group_query_slot_mapping_buffers[primary_gid][:num_query_total]
-        cad.positions = self.positions[:num_query_total]
-        cad.causal = False
-        cad.attn_mask = None
-        cad.attn_state = AscendAttentionState.ChunkedPrefill
-        self._dummy_query_start_loc = cad.query_start_loc_cpu[: batch_size + 1]
-        self._dummy_seq_lens = cad.seq_lens[:batch_size]
-
-        return num_query_total, token_indices_to_sample, cad, None
-
-
-    @torch.inference_mode()
-    def dummy_run(
-        self,
-        num_tokens: int,
-        num_reqs: int = 0,
-        num_tokens_across_dp: torch.Tensor | None = None,
-        aclgraph_runtime_mode: CUDAGraphMode = CUDAGraphMode.NONE,
-        batch_descriptor=None,
-        dummy_compute_logits=lambda hidden_states: None,
-        is_profile=False,
-        **kwargs,
-    ) -> None:
-        # Run dummy_run at full load: the query length of each request is self.num_speculative_tokens
-        # Unlike DFlash, where the query length is self.num_speculative_tokens + 1.
-        # Ensure that the maximum batch token is within the limit of self.max_query_tokens.
-        num_query_per_req = self.num_speculative_tokens
-        num_query_total = num_reqs * num_query_per_req
-        num_query_tokens = min(num_query_total if num_reqs > 0 else num_tokens, self.max_query_tokens)
-
-        (
-            num_input_tokens,
-            num_tokens_across_dp,
-            _,
-        ) = self.runner._sync_metadata_across_dp(num_query_tokens, is_draft_model=True)
-
-        if not self.use_cuda_graph:
-            aclgraph_runtime_mode = CUDAGraphMode.NONE
-
-        context_positions = self._context_positions_buffer[:num_input_tokens]
-        context_states = self.hidden_states[:num_input_tokens]
-
-        self.token_indices_to_sample.fill_(0)
-        self._pad_draft_buffers(num_query_total, num_input_tokens)
-
-        with set_ascend_forward_context(
-            None,
-            self.vllm_config,
-            num_tokens=num_input_tokens,
-            num_tokens_across_dp=num_tokens_across_dp,
-            num_actual_tokens=num_input_tokens,
-            in_profile_run=is_profile,
-            batch_descriptor=batch_descriptor,
-            aclgraph_runtime_mode=aclgraph_runtime_mode,
-            is_draft_model=True,
-            draft_attn_metadatas=[],
-        ):
-            if is_profile:
-                self.model.precompute_and_store_context_kv(context_states, context_positions)
-                self.model(
-                    input_ids=self.input_ids[:num_query_total],
-                    positions=self._get_positions(num_query_total),
-                    inputs_embeds=None,
-                )
-
-            else:
-                self._dflash_num_context = num_input_tokens
-                self._runnable(
-                    num_input_tokens=num_input_tokens,
-                    batch_size=num_reqs,
-                    token_indices_to_sample=self.token_indices_to_sample[: num_reqs * self.num_speculative_tokens],
-                    target_positions=self._get_positions(num_input_tokens),
-                    inputs_embeds=None,
-                    multi_steps_attn_metadata=[],
-                    num_tokens=num_input_tokens,
-                )
